requirements.txt
azure-ai-documentintelligence==1.0.1
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.35.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
isodate==0.7.2
msrest==0.7.1
oauthlib==3.3.1
python-dotenv==1.0.1
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
